Feat: Authorization

+++ app/Http/Controllers/Api/EventController.php
@@ -7,6 +7,7 @@ use App\Http\Resources\EventResource;
 use App\Http\Traits\CanLoadRelationship;
 use App\Models\Event;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class EventController extends Controller
 {
@@ -17,7 +18,6 @@ class EventController extends Controller
 
     public function index()
     {
-        // $query = $this->loadRelationships(Event::query(), $relations);
         $query = $this->loadRelationships(Event::query(), $this->relations);
 
         return EventResource::collection(
@@ -32,6 +32,8 @@ class EventController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('create', Event::class);
+
         $data = $request->validate([
             'name' => 'required|string|max:255',
             'description' => 'nullable|string',
@@ -42,7 +44,7 @@ class EventController extends Controller
 
         $event = Event::create([
             ...$data,
-            'user_id' => 1
+            'user_id' => $request->user()->id
         ]);
 
         return new EventResource($this->loadRelationships($event));
@@ -53,7 +55,6 @@ class EventController extends Controller
      */
     public function show(Event $event)
     {
-        // $event->load('user', 'attendees');
         return new EventResource($this->loadRelationships($event));
     }
 
@@ -62,6 +63,8 @@ class EventController extends Controller
      */
     public function update(Request $request, Event $event)
     {
+        Gate::authorize('update', Event::class);
+        
         $event->update($request->validate([
             'name' => 'sometimes|string|max:255',
             'description' => 'nullable|string',
@@ -77,6 +80,7 @@ class EventController extends Controller
      */
     public function destroy(Event $event)
     {
+        Gate::authorize('delete', Event::class);
         $event->delete();
 
         return response()->json(status: 204);
